+++ ui/main/game/galactic_war/cards/gwc_energy_efficiency_all.js
@@ -89,7 +89,6 @@ define([
         gwaioUnits.sparkWeapon,
         gwaioUnits.sxxWeapon,
         gwaioUnits.artemisWeapon,
-        gwaioUnits.heliosWeaponAG,
       ];
       weapons.forEach(function (weapon) {
         mods.push(