+++ ui/main/game/galactic_war/cards/gwc_energy_efficiency_weapons.js
@@ -34,7 +34,6 @@ define([
         gwaioUnits.sparkWeapon,
         gwaioUnits.sxxWeapon,
         gwaioUnits.artemisWeapon,
-        gwaioUnits.heliosWeaponAG,
       ];
       var mods = [];
       weapons.forEach(function (weapon) {