+++ ui/mods/com.pa.quitch.gwaioverhaul/shared/units.js
@@ -162,9 +162,7 @@ define({
   halley: "/pa/units/orbital/delta_v_engine/delta_v_engine.json",
   helios: "/pa/units/orbital/titan_orbital/titan_orbital.json",
   heliosAmmo: "/pa/units/orbital/titan_orbital/titan_orbital_ammo.json",
-  heliosWeaponAG:
-    "/pa/units/orbital/titan_orbital/titan_orbital_tool_weapon_ground.json",
-  heliosWeaponAO:
+  heliosWeapon:
     "/pa/units/orbital/titan_orbital/titan_orbital_tool_weapon_orbital.json",
   hermes: "/pa/units/orbital/orbital_probe/orbital_probe.json",
   holkins: "/pa/units/land/artillery_long/artillery_long.json",